+++ public/components/Guide.jsx
@@ -22,6 +22,7 @@ import GuideShortcut from "./guide/Shortcut";
 import GuideTable from "./guide/Table";
 import GuideMemoryTable from "./guide/MemoryTable";
 import GuidePagination from "./guide/Pagination";
+import GuideBreadcrumbs from "./guide/Breadcrumbs";
 
 export default class Guide extends React.Component {
     constructor(props){
@@ -84,6 +85,8 @@ export default class Guide extends React.Component {
                                 <Route path='/guide/action-button' component={GuideActionButton}/>
                                 <Route path='/guide/shortcut' component={GuideShortcut}/>
                                 <Route path='/guide/pagination' component={GuidePagination}/>
+                                <Route path='/guide/breadcrumbs' component={GuideBreadcrumbs}/>
+
                                 <Route component={NotFound} />
                             </Switch>
                         </article>