+++ public/components/guide/Template.jsx
@@ -1,11 +1,15 @@
 import React from "react";
-import {Adsense} from "../../../src";
+import {Adsense, Table} from "../../../src";
 import PrismCode from "../PrismCode";
 import Example from "../Example";
+import {tablePropsHeader} from "../Defs";
 
 export default class GuideTemplate extends React.Component {
     render(){
         const codeImport = `import {Component} from "metro4-react";`;
+        const codeUsing = ``;
+
+        const tablePropsBody = [];
 
         return(
             <div>
@@ -28,6 +32,14 @@ export default class GuideTemplate extends React.Component {
                 <br/>
                 <h3>Importing</h3>
                 <PrismCode language="js" code={codeImport}/>
+
+                <br/>
+                <h3>Using</h3>
+                <PrismCode language="js" code={codeUsing}/>
+
+                <br/>
+                <h4>Props</h4>
+                <Table className="table-border cell-border" head={tablePropsHeader} body={tablePropsBody}/>
 
                 <br/>
                 <Adsense client={'ca-pub-1632668592742327'} slot={'4639163605'} test={SITE_MODE_DEV}/>