+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ### 0.1.1
 + [x] Container: minor updates for using additional props
++ [x] Breadcrumbs: updates for using props, add docs
 
 ### 0.1.0
 + [x] Initial version, contains 40 components, common styles, additional colors, routines.